requirements.txt
Django==1.9.5
django-axes==1.6.1
django-ckeditor==5.0.3
django-filer==1.2.0
django-mptt==0.8.4
django-polymorphic==0.8.1
django-tinymce==2.3.0
djangorestframework==3.3.3
djangorestframework-csv==1.4.1
easy-thumbnails==2.3
loremipsum==1.0.5
Markdown==2.6.6
mysqlclient==1.3.6
Pillow==3.2.0
pytz==2015.7
six==1.10.0
Unidecode==0.4.19
psycopg2==2.6.1
